+++ tests/input/unused-import/_fixture_for_conftest.py
@@ -1,0 +1,6 @@
+import pytest
+
+
+@pytest.fixture
+def conftest_fixture_attr():
+    return True